evolve/requirements.txt
psycopg2-binary
pika
numpy
polars==0.16.11
pydantic==1.10.11
python-dotenv
cerberus
timezonefinder
pvlib
matplotlib
pyarrow